requirements.txt
acme==1.12.0
anyio==3.6.2
arandr==0.1.10
astroid==2.5.1
asttokens==2.0.4
automationhat==0.2.0
beautifulsoup4==4.9.3
blinker==1.4
blinkt==0.1.2
buttonshim==0.0.2
Cap1xxx==0.1.3
certbot==1.12.0
certifi==2020.6.20
chardet==4.0.0
click==7.1.2
colorama==0.4.4
colorzero==1.1
ConfigArgParse==1.2.3
configobj==5.0.6
cryptography==3.3.2
cupshelpers==1.0
dbus-python==1.2.16
distro==1.5.0
dnspython==2.2.1
docutils==0.16
dotenv==0.0.5
drumhat==0.1.0
email-validator==1.3.0
envirophat==1.0.0
ExplorerHAT==0.4.2
fastapi==0.88.0
Flask==1.1.2
fourletterphat==0.1.0
gpiozero==1.6.2
h11==0.14.0
html5lib==1.1
httpcore==0.16.2
httptools==0.5.0
httpx==0.23.1
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
josepy==1.2.0
lazy-object-proxy==0.0.0
logilab-common==1.8.1
lxml==4.6.3
MarkupSafe==1.1.1
mccabe==0.6.1
microdotphat==0.2.1
mote==0.0.4
motephat==0.0.3
mypy==0.812
mypy-extensions==0.4.3
numpy==1.19.5
oauthlib==3.1.0
orjson==3.8.2
pantilthat==0.0.7
parsedatetime==2.6
parso==0.8.1
pexpect==4.8.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
picamera2==0.3.6
pidng==4.0.9
piexif==1.1.3
piglow==1.2.5
pigpio==1.78
Pillow==8.1.2
psutil==5.8.0
pycairo==1.16.2
pycups==2.0.1
pydantic==1.10.2
pygame==1.9.6
Pygments==2.7.1
PyGObject==3.38.0
PyICU==2.5
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.7.2
PyOpenGL==3.1.5
pyOpenSSL==20.0.1
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyRFC3339==1.1
pyserial==3.5b0
pysmbc==1.0.23
python-apt==2.2.1
python-dotenv==0.21.0
python-multipart==0.0.5
python-prctl==1.7
pytz==2021.1
PyYAML==6.0
rainbowhat==0.1.0
reportlab==3.5.59
requests==2.25.1
requests-oauthlib==1.0.0
requests-toolbelt==0.9.1
responses==0.12.1
rfc3986==1.5.0
roman==2.0.0
RPi.GPIO==0.7.0
RTIMULib==7.2.1
scrollphat==0.0.7
scrollphathd==1.2.1
Send2Trash==1.6.0b1
sense-hat==2.4.0
simplejpeg==1.6.4
simplejson==3.17.2
six==1.16.0
skywriter==0.0.7
sn3218==1.2.7
sniffio==1.3.0
soupsieve==2.2.1
spidev==3.5
ssh-import-id==5.10
starlette==0.22.0
thonny==3.3.14
toml==0.10.1
touchphat==0.0.1
twython==3.8.2
typed-ast==1.4.2
typing-extensions==4.4.0
ufw==0.36
ujson==5.5.0
unicornhathd==0.0.4
urllib3==1.26.5
uvicorn==0.20.0
uvloop==0.17.0
v4l2-python3==0.3.1
watchfiles==0.18.1
webencodings==0.5.1
websockets==10.4
Werkzeug==1.0.1
wrapt==1.12.1
wsproto==0.15.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.0.1
zope.interface==5.2.0
